refactor(scripts): file-organization sweep over dependencyGraph
The scripts tree takes the app's layer-first layout: the command's functions and constants move to services/dependencyGraph, its types to models/dependencyGraph, and the entrypoint folder keeps index.ts alone. The graph freshness test reads the repository root from REPOSITORY_ROOT rather than counting its own depth, which the move would have put one short. The rule lands in the file-organization skill, the recipe and catalogue references and the scripts README with it.
