+++ .agents/skills/file-organization/SKILL.md
@@ -30,6 +30,7 @@ description: Esposter file and folder organisation — the alias imports (shared
 - **External library extensions go in `services/`** — helpers that extend/wrap third-party libraries (`services/<lib>/doThing.ts`), not `util/`. A helper over a language global rather than a library is a `util/` one (`util/date/formatDate.ts`).
 - **`utils/` is for truly universal utilities only** — math, string, regex, type utilities, Node/browser engine extensions with no external dependency. If the helper imports a third-party package, it belongs in `services/`. Generic browser utilities go in `app/utils/` (e.g. `readFoo.ts`).
 - **Feature folders** — group related models/services/components under a feature subfolder (e.g. `feature/sub-feature/`).
+- **`scripts/` is laid out by layer like the app, keyed by command** — `src/<command>/index.ts` is the entrypoint a pnpm script names (an oxlint plugin is `src/oxlint/<plugin>.ts`, the path `.oxlintrc.json` loads), and everything it runs lives under `src/services/<command>/` (its `constants.ts` included) with its types under `src/models/<command>/`. A sub-command (`coderabbit/<verb>`, `sweeps/<scan>`) nests one level deeper in each; what two verbs share sits at their command's root in the layer. The entrypoint folder holds nothing but the entrypoint and the platform scripts it dispatches to.
 - **Sole-consumer subfolder rule (CRITICAL).** A file lives in the subfolder of the **one feature that consumes it**, and stays at the parent level only when two or more sibling features import it or it is a shared primitive with no owning feature — which is what keeps a directory from accumulating twenty loose files from distinct sub-concerns. A `models/` folder mirroring a `services/` folder mirrors its feature subfolders too, leaving only cross-feature types at the root. Don't over-fragment the other way: a shared-primitive bucket stays whole even when large, and an already-feature-organised folder is not nested further.
 - **No magic strings** — always use enums for discriminants, command types, and other categorical values. Before typing any literal, search the repo for a value that already means it and import that: an enum member (`Operation.Read`, `DatabaseEntityType.Post`), a separator (`ID_SEPARATOR`), a registry entry (`AsyncDataKey`, `LocalStorageKey`, `RoutePath`), a mime type off the configuration map. A literal is earned only when nothing existing means it — a second spelling of something the repo already names never renames with the original.
