+++ .agents/skills/package-scripts/references/ai-scripts.md
@@ -4,8 +4,8 @@ Read when a sweep, a review or a skill needs the script that runs its scan — e
 entry, what it prints, and what it costs. The naming rule is the `skill-authoring` skill's (`references/embedded-recipes.md`);
 this page is the list.
 
-Each is a `pnpm -C scripts run <same name>` delegation to `scripts`, which owns the source under
-`scripts/src/<domain>/<verb>/` with a colocated test. One `ai:sweep:*` per scan.
+Each is a `pnpm -C scripts run <same name>` delegation to `scripts`, whose entrypoint is
+`scripts/src/<domain>/<verb>/index.ts` and whose functions sit under `scripts/src/services/<domain>/<verb>/` with a colocated test. One `ai:sweep:*` per scan.
 
 | Command                                   | Prints                                                                                                                                                                                                                                                                                     |
 | ----------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
