+++ scripts/src/dependencyGraph/index.ts
@@ -1,7 +1,7 @@
-import { GRAPH_FILENAME } from "#src/dependencyGraph/constants";
-import { getGraphSvg } from "#src/dependencyGraph/getGraphSvg";
-import { getWorkspaceEdges } from "#src/dependencyGraph/getWorkspaceEdges";
 import { REPOSITORY_ROOT } from "#src/services/constants";
+import { GRAPH_FILENAME } from "#src/services/dependencyGraph/constants";
+import { getGraphSvg } from "#src/services/dependencyGraph/getGraphSvg";
+import { getWorkspaceEdges } from "#src/services/dependencyGraph/getWorkspaceEdges";
 import { getWorkspacePackages } from "#src/services/getWorkspacePackages";
 import { writeFileSync } from "node:fs";
 import { resolve } from "node:path";
