+++ scripts/src/models/dependencyGraph/WorkspaceEdge.ts
