+++ scripts/src/models/dependencyGraph/WorkspaceEdges.ts
@@ -1,4 +1,4 @@
-import type { WorkspaceEdge } from "#src/dependencyGraph/models/WorkspaceEdge";
+import type { WorkspaceEdge } from "#src/models/dependencyGraph/WorkspaceEdge";
 
 export interface WorkspaceEdges {
   development: WorkspaceEdge[];
