+++ scripts/src/services/dependencyGraph/getGraphSvg.test.ts
@@ -1,6 +1,7 @@
-import { GRAPH_FILENAME } from "#src/dependencyGraph/constants";
-import { getGraphSvg } from "#src/dependencyGraph/getGraphSvg";
-import { getWorkspaceEdges } from "#src/dependencyGraph/getWorkspaceEdges";
+import { REPOSITORY_ROOT } from "#src/services/constants";
+import { GRAPH_FILENAME } from "#src/services/dependencyGraph/constants";
+import { getGraphSvg } from "#src/services/dependencyGraph/getGraphSvg";
+import { getWorkspaceEdges } from "#src/services/dependencyGraph/getWorkspaceEdges";
 import { getWorkspacePackages } from "#src/services/getWorkspacePackages";
 import { readFileSync } from "node:fs";
 import { resolve } from "node:path";
@@ -14,10 +15,9 @@ describe(getGraphSvg, () => {
   test("matches the committed graph", async () => {
     expect.hasAssertions();
 
-    const root = resolve(import.meta.dirname, "..", "..", "..");
-    const workspacePackages = getWorkspacePackages(root);
+    const workspacePackages = getWorkspacePackages(REPOSITORY_ROOT);
     await expect(getGraphSvg(workspacePackages, getWorkspaceEdges(workspacePackages))).resolves.toBe(
-      readFileSync(resolve(root, GRAPH_FILENAME), "utf8"),
+      readFileSync(resolve(REPOSITORY_ROOT, GRAPH_FILENAME), "utf8"),
     );
   });
 });
